HBASE-2054 memstore size 0 is >= than blocking -2.0g size
git-svn-id: https://svn.apache.org/repos/asf/hadoop/hbase/trunk@892653 13f79535-47bb-0310-9956-ffa450edef68

+++ src/java/org/apache/hadoop/hbase/HTableDescriptor.java
@@ -424,7 +424,7 @@ public class HTableDescriptor implements WritableComparable<HTableDescriptor> {
   public long getMemStoreFlushSize() {
     byte [] value = getValue(MEMSTORE_FLUSHSIZE_KEY);
     if (value != null)
-      return Integer.valueOf(Bytes.toString(value)).intValue();
+      return Long.valueOf(Bytes.toString(value)).longValue();
     return DEFAULT_MEMSTORE_FLUSH_SIZE;
   }
